Fix: The number of ports on the XHCI is configured through boot parameters.
The XHCI has only two ports, whitch limits the number of USB devices.
Now it is configred by boot parameter. The maximum number cannot exceed
15. If this parameter is not set, the default value is 4.

+++ usb/src/usb.rs
@@ -141,11 +141,11 @@ pub struct UsbPort {
     pub dev: Option<Arc<Mutex<dyn UsbDeviceOps>>>,
     pub speed_mask: u32,
     pub path: String,
-    pub index: u32,
+    pub index: u8,
 }
 
 impl UsbPort {
-    pub fn new(index: u32) -> Self {
+    pub fn new(index: u8) -> Self {
         Self {
             dev: None,
             speed_mask: 0,
